requirements.txt
aiofiles==0.8.0
aiosqlite==0.17.0
anyio==3.5.0
appdirs==1.4.4
async-timeout==4.0.2
asyncpg==0.25.0
atomicwrites==1.4.0
attrs==20.3.0
bcrypt==3.2.0
boto3==1.20.35
botocore==1.23.54
CacheControl==0.12.11
cachetools==5.0.0
certifi==2020.12.5
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==7.1.2
colorama==0.4.4
decorator==5.1.1
Deprecated==1.2.13
Faker==13.5.0
fastapi==0.62.0
firebase==3.0.1
firebase-admin==5.2.0
fireorm==0.0.14
fs==2.4.15
future==0.18.2
gcloud-storage-emulator==0.5.0
geocoder==1.38.1
google-api-core==2.7.2
google-api-python-client==2.45.0
google-auth==2.6.5
google-auth-httplib2==0.1.0
google-cloud-core==2.3.0
google-cloud-datastore==1.15.4
google-cloud-firestore==2.4.0
google-cloud-ndb==1.11.1
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
grpcio==1.44.0
grpcio-status==1.44.0
gunicorn==20.1.0
h11==0.11.0
httpcore==0.14.7
httplib2==0.20.4
httptools==0.4.0
httpx==0.22.0
idna==2.10
iniconfig==1.1.1
iso8601==0.1.16
jmespath==0.10.0
mangum==0.10.0
msgpack==1.0.3
packaging==20.8
passlib==1.7.4
Pillow==9.0.0
pluggy==0.13.1
proto-plus==1.20.3
protobuf==3.20.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.7.3
PyJWT==2.3.0
pymemcache==3.5.2
pyparsing==2.4.7
pypika-tortoise==0.1.4
pytest==6.2.0
pytest-mock==3.7.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2022.1
qrcode==7.3.1
ratelim==0.1.6
redis==4.2.2
requests==2.25.0
rfc3986==1.5.0
rsa==4.8
s3transfer==0.5.2
six==1.16.0
sniffio==1.2.0
starlette==0.13.6
toml==0.10.2
tortoise-orm==0.18.1
typing-extensions==3.7.4.3
uritemplate==4.1.1
urllib3==1.26.2
uvicorn==0.13.1
uvloop==0.16.0
wrapt==1.14.0
